feat: update pagination for lazy fetch of curius
