+++ quartz/components/scripts/curius.inline.ts
@@ -1,188 +1,20 @@
-import { registerEscapeHandler, removeAllChildren, registerEvents } from "./util"
 import { Link } from "../types"
 import {
   fetchCuriusLinks,
-  createTitle,
-  timeSince,
+  fetchSearchLinks,
   createTrailMetadata,
   createTrailList,
   curiusSearch,
+  createLinkEl,
 } from "./curius"
+import { updateNavigation } from "./curius-navigation.inline"
 
-let currentActive: HTMLLIElement | null = null
-function createLinkEl(Link: Link): HTMLLIElement {
-  const curiusItem = document.createElement("li")
-  curiusItem.id = `curius-item-${Link.id}`
-  curiusItem.classList.add("curius-item")
-
-  const createMetadata = (Link: Link): HTMLDivElement => {
-    const item = document.createElement("div")
-    item.classList.add("curius-item-metadata")
-
-    const tags = document.createElement("ul")
-    tags.classList.add("curius-item-tags")
-    tags.innerHTML =
-      Link.topics.length > 0
-        ? `${Link.topics
-            .map((topic) =>
-              topic.public
-                ? `<li><a href="https://curius.app/aaron-pham/${topic.slug}" target="_blank">${topic.topic}</a></li>`
-                : ``,
-            )
-            .join("")}`
-        : ``
-
-    const misc = document.createElement("div")
-    misc.id = `curius-misc-${Link.id}`
-    const time = document.createElement("span")
-    time.id = `curius-span-${Link.id}`
-    const modifiedDate = new Date(Link.modifiedDate as string)
-    time.innerHTML = `<time datetime=${
-      Link.modifiedDate
-    } title="${modifiedDate.toUTCString()}">${timeSince(Link.createdDate as string)}</time>`
-    misc.appendChild(time)
-
-    if (Link.highlights.length > 0) {
-      const highlights = document.createElement("div")
-      highlights.id = `curius-highlights-${Link.id}`
-      highlights.innerHTML = `${Link.highlights.length} ${Link.highlights.length > 0 ? "highlights" : "highlight"}`
-      misc.appendChild(highlights)
-
-      const modal = document.getElementById("highlight-modal")
-      const modalList = document.getElementById("highlight-modal-list")
-
-      const onMouseEnter = () => {
-        const highlightsData = Link.highlights
-
-        if (!modal || !modalList) return
-        // clear the previous modal
-        modalList.innerHTML = ""
-        curiusItem.classList.remove("focus")
-
-        highlightsData.forEach((highlight) => {
-          let hiItem = document.createElement("li")
-          hiItem.textContent = highlight.highlight
-          modalList.appendChild(hiItem)
-        })
-        modal.style.visibility = "visible"
-        modal.classList.add("active")
-      }
-
-      const onMouseLeave = () => {
-        curiusItem.classList.add("focus")
-
-        if (!modal) return
-        modal.style.visibility = "hidden"
-        modal.classList.remove("active")
-      }
-
-      const onMouseMove = ({ pageX, pageY }: MouseEvent) => {
-        curiusItem.classList.remove("focus")
-
-        if (!modal) return
-        modal.classList.add("active")
-        modal.style.left = `${pageX + 10}px`
-        modal.style.top = `${pageY + 10}px`
-      }
-
-      registerEvents(
-        highlights,
-        ["mouseenter", onMouseEnter],
-        ["mouseleave", onMouseLeave],
-        ["mousemove", onMouseMove],
-      )
+declare global {
+  interface Window {
+    curiusState?: {
+      currentPage: number
+      hasMore: boolean
     }
-
-    item.append(tags, misc)
-    return item
-  }
-
-  curiusItem.append(createTitle({ Link, addFaIcon: true }), createMetadata(Link))
-  curiusItem.dataset.items = JSON.stringify(true)
-
-  const onClick = (e: HTMLElementEventMap["click"]) => {
-    const note = document.getElementsByClassName("curius-notes")[0] as HTMLDivElement | null
-
-    if (e.altKey || e.ctrlKey || e.metaKey || e.shiftKey) return
-    if (currentActive) currentActive.classList.remove("active")
-    if (note) note.classList.remove("active")
-
-    currentActive = curiusItem
-    currentActive.classList.add("active")
-
-    if (Link.highlights.length > 0) {
-      if (!note) return
-      note.classList.add("active")
-      updateNotePanel(Link, note, currentActive)
-    }
-
-    if (e.target instanceof HTMLAnchorElement || note?.classList.contains("active")) return
-    window.open(Link.link, "_blank")
-  }
-
-  registerEscapeHandler(curiusItem, () => curiusItem.classList.remove("active"))
-
-  const onMouseEnter = () => {
-    const favoriteDiv = curiusItem.querySelector("svg.favorite-icon") as HTMLDivElement | null
-
-    if (favoriteDiv) favoriteDiv.classList.add("focus")
-    curiusItem.classList.add("focus")
-  }
-
-  const onMouseLeave = () => {
-    const favoriteDiv = curiusItem.querySelector("svg.favorite-icon") as HTMLDivElement | null
-
-    if (favoriteDiv) favoriteDiv.classList.remove("focus")
-    curiusItem.classList.remove("focus")
-  }
-
-  registerEvents(
-    curiusItem,
-    //@ts-ignore
-    ["click", onClick],
-    ["mouseenter", onMouseEnter],
-    ["mouseleave", onMouseLeave],
-  )
-
-  return curiusItem
-}
-
-function updateNotePanel(Link: Link, note: HTMLDivElement, parent: HTMLLIElement) {
-  const titleNode = note.querySelector("#note-link") as HTMLAnchorElement
-  const snippetNode = note.querySelector(".curius-note-snippet") as HTMLDivElement
-  const highlightsNode = note.querySelector(".curius-note-highlights") as HTMLDivElement
-
-  titleNode.innerHTML = `<span class="curius-item-span">${Link.title}</span>`
-  titleNode.href = Link.link
-  titleNode.target = "_blank"
-  titleNode.rel = "noopener noreferrer"
-
-  const close = document.querySelector(".icon-container")
-
-  const cleanUp = () => {
-    note.classList.remove("active")
-    parent.classList.remove("active")
-  }
-
-  close?.addEventListener("click", cleanUp)
-  window.addCleanup(() => close?.removeEventListener("click", cleanUp))
-  registerEscapeHandler(note, cleanUp)
-
-  removeAllChildren(snippetNode)
-  snippetNode.textContent = Link.metadata ? Link.metadata.full_text : Link.snippet
-
-  removeAllChildren(highlightsNode)
-  if (Link.highlights.length === 0) return
-  for (const hl of Link.highlights) {
-    const highlightItem = document.createElement("li")
-    const hlLink = document.createElement("a")
-    hlLink.dataset.highlight = hl.id.toString()
-    hlLink.href = `${Link.link}?curius=${hl.userId}`
-    hlLink.target = "_blank"
-    hlLink.rel = "noopener noreferrer"
-    hlLink.textContent = hl.highlight
-    highlightItem.appendChild(hlLink)
-    highlightsNode.appendChild(highlightItem)
   }
 }
 
@@ -202,7 +34,9 @@ document.addEventListener("nav", async (e) => {
 
   fetchText.textContent = "Récupération des liens curius"
   fetchText.classList.toggle("active", true)
-  const resp = await fetchCuriusLinks()
+
+  const [resp, searchLinks] = await Promise.all([fetchCuriusLinks(), fetchSearchLinks()])
+
   fetchText.classList.toggle("active", false)
 
   const callIfEmpty = (data: Link[]) => {
@@ -213,15 +47,24 @@ document.addEventListener("nav", async (e) => {
     return data.filter((link) => link.trails.length === 0)
   }
 
-  // show trails separately
   const linksData = callIfEmpty(resp.links!)
   if (linksData.length === 0) return
 
   createTrailList(createTrailMetadata(resp))
 
-  await curiusSearch(linksData)
+  await curiusSearch(searchLinks)
 
   fragment.append(...linksData.map(createLinkEl))
+
   if (friends) friends.classList.toggle("active", true)
   if (trails) trails.classList.toggle("active", true)
+
+  // Store pagination state
+  window.curiusState = {
+    currentPage: 0,
+    hasMore: resp.hasMore ?? false,
+  }
+
+  // Update navigation buttons
+  updateNavigation()
 })
