+++ quartz/components/scripts/curius.ts
@@ -55,6 +55,184 @@ function extractApexDomain(url: string) {
   return match ? match[1] : ""
 }
 
+let currentActive: HTMLLIElement | null = null
+
+function updateNotePanel(Link: Link, note: HTMLDivElement, parent: HTMLLIElement) {
+  const titleNode = note.querySelector("#note-link") as HTMLAnchorElement
+  const snippetNode = note.querySelector(".curius-note-snippet") as HTMLDivElement
+  const highlightsNode = note.querySelector(".curius-note-highlights") as HTMLDivElement
+
+  titleNode.innerHTML = `<span class="curius-item-span">${Link.title}</span>`
+  titleNode.href = Link.link
+  titleNode.target = "_blank"
+  titleNode.rel = "noopener noreferrer"
+
+  const close = document.querySelector(".icon-container")
+
+  const cleanUp = () => {
+    note.classList.remove("active")
+    parent.classList.remove("active")
+  }
+
+  close?.addEventListener("click", cleanUp)
+  window.addCleanup(() => close?.removeEventListener("click", cleanUp))
+  registerEscapeHandler(note, cleanUp)
+
+  removeAllChildren(snippetNode)
+  snippetNode.textContent = Link.metadata ? Link.metadata.full_text : Link.snippet
+
+  removeAllChildren(highlightsNode)
+  if (Link.highlights.length === 0) return
+  for (const hl of Link.highlights) {
+    const highlightItem = document.createElement("li")
+    const hlLink = document.createElement("a")
+    hlLink.dataset.highlight = hl.id.toString()
+    hlLink.href = `${Link.link}?curius=${hl.userId}`
+    hlLink.target = "_blank"
+    hlLink.rel = "noopener noreferrer"
+    hlLink.textContent = hl.highlight
+    highlightItem.appendChild(hlLink)
+    highlightsNode.appendChild(highlightItem)
+  }
+}
+
+export function createLinkEl(Link: Link): HTMLLIElement {
+  const curiusItem = document.createElement("li")
+  curiusItem.id = `curius-item-${Link.id}`
+  curiusItem.classList.add("curius-item")
+
+  const createMetadata = (Link: Link): HTMLDivElement => {
+    const item = document.createElement("div")
+    item.classList.add("curius-item-metadata")
+
+    const tags = document.createElement("ul")
+    tags.classList.add("curius-item-tags")
+    tags.innerHTML =
+      Link.topics.length > 0
+        ? `${Link.topics
+            .map((topic) =>
+              topic.public
+                ? `<li><a href="https://curius.app/aaron-pham/${topic.slug}" target="_blank">${topic.topic}</a></li>`
+                : ``,
+            )
+            .join("")}`
+        : ``
+
+    const misc = document.createElement("div")
+    misc.id = `curius-misc-${Link.id}`
+    const time = document.createElement("span")
+    time.id = `curius-span-${Link.id}`
+    const modifiedDate = new Date(Link.modifiedDate as string)
+    time.innerHTML = `<time datetime=${
+      Link.modifiedDate
+    } title="${modifiedDate.toUTCString()}">${timeSince(Link.createdDate as string)}</time>`
+    misc.appendChild(time)
+
+    if (Link.highlights.length > 0) {
+      const highlights = document.createElement("div")
+      highlights.id = `curius-highlights-${Link.id}`
+      highlights.innerHTML = `${Link.highlights.length} ${Link.highlights.length > 0 ? "highlights" : "highlight"}`
+      misc.appendChild(highlights)
+
+      const modal = document.getElementById("highlight-modal")
+      const modalList = document.getElementById("highlight-modal-list")
+
+      const onMouseEnter = () => {
+        const highlightsData = Link.highlights
+
+        if (!modal || !modalList) return
+        // clear the previous modal
+        modalList.innerHTML = ""
+        curiusItem.classList.remove("focus")
+
+        highlightsData.forEach((highlight) => {
+          let hiItem = document.createElement("li")
+          hiItem.textContent = highlight.highlight
+          modalList.appendChild(hiItem)
+        })
+        modal.style.visibility = "visible"
+        modal.classList.add("active")
+      }
+
+      const onMouseLeave = () => {
+        curiusItem.classList.add("focus")
+
+        if (!modal) return
+        modal.style.visibility = "hidden"
+        modal.classList.remove("active")
+      }
+
+      const onMouseMove = ({ pageX, pageY }: MouseEvent) => {
+        curiusItem.classList.remove("focus")
+
+        if (!modal) return
+        modal.classList.add("active")
+        modal.style.left = `${pageX + 10}px`
+        modal.style.top = `${pageY + 10}px`
+      }
+
+      registerEvents(
+        highlights,
+        ["mouseenter", onMouseEnter],
+        ["mouseleave", onMouseLeave],
+        ["mousemove", onMouseMove],
+      )
+    }
+
+    item.append(tags, misc)
+    return item
+  }
+
+  curiusItem.append(createTitle({ Link, addFaIcon: true }), createMetadata(Link))
+  curiusItem.dataset.items = JSON.stringify(true)
+
+  const onClick = (e: HTMLElementEventMap["click"]) => {
+    const note = document.getElementsByClassName("curius-notes")[0] as HTMLDivElement | null
+
+    if (e.altKey || e.ctrlKey || e.metaKey || e.shiftKey) return
+    if (currentActive) currentActive.classList.remove("active")
+    if (note) note.classList.remove("active")
+
+    currentActive = curiusItem
+    currentActive.classList.add("active")
+
+    if (Link.highlights.length > 0) {
+      if (!note) return
+      note.classList.add("active")
+      updateNotePanel(Link, note, currentActive)
+    }
+
+    if (e.target instanceof HTMLAnchorElement || note?.classList.contains("active")) return
+    window.open(Link.link, "_blank")
+  }
+
+  registerEscapeHandler(curiusItem, () => curiusItem.classList.remove("active"))
+
+  const onMouseEnter = () => {
+    const favoriteDiv = curiusItem.querySelector("svg.favorite-icon") as HTMLDivElement | null
+
+    if (favoriteDiv) favoriteDiv.classList.add("focus")
+    curiusItem.classList.add("focus")
+  }
+
+  const onMouseLeave = () => {
+    const favoriteDiv = curiusItem.querySelector("svg.favorite-icon") as HTMLDivElement | null
+
+    if (favoriteDiv) favoriteDiv.classList.remove("focus")
+    curiusItem.classList.remove("focus")
+  }
+
+  registerEvents(
+    curiusItem,
+    //@ts-ignore
+    ["click", onClick],
+    ["mouseenter", onMouseEnter],
+    ["mouseleave", onMouseLeave],
+  )
+
+  return curiusItem
+}
+
 export function timeSince(date: Date | string) {
   const now = new Date()
   const dateObject = date instanceof Date ? date : new Date(date)
@@ -342,10 +520,6 @@ let index = new FlexSearch.Document({
         field: "snippet",
         tokenize: "forward",
       },
-      {
-        field: "trails",
-        tokenize: "forward",
-      },
     ],
   },
 })
@@ -362,7 +536,7 @@ const tokenizeTerm = (term: string) => {
     }
   }
 
-  return tokens.sort((a, b) => b.length - a.length) // always highlight longest terms first
+  return tokens.sort((a, b) => b.length - a.length)
 }
 
 function highlight(searchTerm: string, text: string, trim?: boolean) {
@@ -394,7 +568,6 @@ function highlight(searchTerm: string, text: string, trim?: boolean) {
 
   const slice = tokenizedText
     .map((tok) => {
-      // see if this tok is prefixed by any search terms
       for (const searchTok of tokenizedTerms) {
         if (tok.toLowerCase().includes(searchTok.toLowerCase())) {
           const regex = new RegExp(searchTok.toLowerCase(), "gi")
@@ -415,13 +588,12 @@ const now = new Date()
 const seed = now.getFullYear() * 10000 + (now.getMonth() + 1) * 100 + now.getDate()
 const rng = new LCG(seed)
 
-export async function curiusSearch(linksData: Link[]) {
-  const sampleLinks = rng.shuffle(linksData).splice(0, 20)
+export async function curiusSearch(searchData: Link[]) {
+  const sampleLinks = rng.shuffle(searchData).splice(0, 20)
 
   const bar = document.getElementById("curius-bar") as HTMLInputElement | null
   const container = document.getElementById("curius-search-container") as HTMLDivElement | null
 
-  // Search functionality
   async function onType(e: HTMLElementEventMap["input"]) {
     let term = (e.target as HTMLInputElement).value
     container?.classList.toggle("active", term !== "")
@@ -429,7 +601,7 @@ export async function curiusSearch(linksData: Link[]) {
       (await index?.searchAsync({
         query: term,
         limit: numSearchResults,
-        index: ["title", "snippet", "topics"],
+        index: ["title", "snippet", "link"],
       })) ?? []
 
     const getByField = (field: string): number[] => {
@@ -440,7 +612,7 @@ export async function curiusSearch(linksData: Link[]) {
     const allIds: Set<number> = new Set([
       ...getByField("title"),
       ...getByField("snippet"),
-      ...getByField("topics"),
+      ...getByField("link"),
     ])
 
     const finalResults = [...allIds].map((id) => formatLinks(term, id))
@@ -448,7 +620,7 @@ export async function curiusSearch(linksData: Link[]) {
   }
 
   const formatLinks = (term: string, id: number): Link => {
-    const L = linksData[id]
+    const L = searchData[id]
     return {
       ...L,
       title: highlight(term, L.title),
@@ -560,7 +732,24 @@ export async function curiusSearch(linksData: Link[]) {
 
   registerEscapeHandler(container, hideLinks)
 
-  await fillIndex(linksData)
+  await fillIndex(searchData)
+}
+
+export async function fetchSearchLinks() {
+  try {
+    const resp = await fetch("https://aarnphm.xyz/api/curius?query=searchLinks", fetchLinksHeaders)
+    if (!resp.ok) {
+      throw new Error("Failed to get searchLinks from curius")
+    }
+    const data: CuriusResponse = await resp.json()
+    if (data === undefined || data.links === undefined) {
+      throw new Error("No searchLinks data")
+    }
+    return data.links
+  } catch (err) {
+    console.error(err)
+    return []
+  }
 }
 
 let indexPopulated = false
